(RE-8727) Add support to define a nonfinal gpg key
This commit adds the ability for users to specify a nonfinal gpg key.
This is an alternative to the main gpg key and will be used to sign any
package versions that are flagged as nonfinal versions. The idea is that
this automation allows us to set up and maintain separate
nonfinal/beta/nightly/testing repos using our main shipping automation
(this repo). This will allow us to get a little bit closer to automating
the release process.

Note, the debian repo configs still specify the main gpg key as the
signing key. This is due to the fact that there was no easy way to
branch on keys for freight. The signing key is actually defined in the
freight config file that lives on the internal repo staging host. We
could add multiple of these config files for each key we sign with. But
it seemed more logical to simply continue signing our apt repos with the
main key.