(maint) Fix a few things
This commit fixes a few typos in the platform tag from path method. It
also updates what architecture we default to when we are dealing with a
noarch/all package. Rather than defaulting to 64bit, we just take the
first architecture in the list of supported architectures for the given
platform.

This also moves fedora and sles above el. This is a brute force fix to
deal with the fact that 'el' is a very simple string to match against,
and matches more than just the OS name, which is a problem if you're
shipping a fedora package and the automation thinks you're dealing with
el