(maint) Add override to make tagged but dirty trees final
When testing automation changes or fixing last-minute packaging errors,
it can be desirable to check out a tagged version of a project and
modify build_defaults to use an alternate branch of packaging. This
allows you to pass the ALLOW_DIRTY_TREE environment variable to return
cause final? to return true if you're on a tagged build with a dirty
tree.