(RE-8979) Deprecate some of the special gem workflows
Shipping gems to the download location don't actually need to be handled
independently from the rest of the shipping code. This commit deprecated
Pkg::Gem.rsync_to_downloads in favor of Pkg::Util::Ship.ship_pkgs. This
also updates the ship_gem_to_downloads task to use the `ship_pkgs`
method, and ship all the gems to downloads in one call rather than
shipping each independently. We still want to ship the gems to
gemservers individually since that doesn't work well with duplicate
publishing, but there's no need to do that for the gems shipping to
downloads.