(maint) Do not shallow clone git repo
Setting `--depth 1` when cloning the build-data repo doesn't serve us
well. It prevents us from checking out the different branches we need to
check out. We need to be able to check out both the team specific and
project specific branch in order to load any extra data from them. We
can't do that with a shallow clone. This fixes that issue.