(maint) Refactor platform_data collection
This commit updates platform_data to make use of tag_from_artifact_path when
mapping platform tags to artifact paths for consumption by beaker. Previously,
we were relying on artifacts_path for the mapping, which makes assumptions about
what we expect to find on builds. Now we take the actual, existing artifact
paths and map to tags.