(RE-8838) Generalize deb repo creation
This commit abstracts and generalizes the logic used to create debian
repos. This isn't perfect, but it means that we are no longer dependent
on a specific path layout to effectively create repos. We can now house
artifacts under whatever directory structure we desire. This is
especially helpful with the switch to puppet5. I'm also taking advantage
of this change to really shake things up and stage built artifacts under
platform tags. The idea is we can easily see all the different
platforms, platform versions, and platform architectures a given sha is
built for. That, and beaker can simply use the logic outlined in the
platform_data hash in the <ref>.yaml file. Nothing should really care
that much how we structure these directories, so I'm probably going to
make peeps angry...