(RE-8838) Stage built packages under platform-tag
This commit completely changes things up and simply drops built packages
under the platform tag for the package. This saves us from having to
care at all about the repo these packages are built for. If we have a
single package that is supposed to be shipped to multiple repos, this
helps with that. This also allows us to make our shipping software
opinionated about the repo path structure, yet another step toward
modularizing out build/package/sign/ship utilities.