(RE-8838) Generalize the paths we search to sign
This commit removes a lot of the assumptions we use when we are signing
rpm packages. We are changing the path layout under which we stage these
artifacts. We should be able to simply query the filesystem, and compile
a list of packages to be signed based on what we find. This commit also
leverages data already available in PLATFORM_INFO so that we no longer
have to hard code and duplicate data in ALL THE PLACES.