(maint) Do not fail if nonfinal_repo_name not set
There are more situations where repo_name is called before we even ship
to the final location. This is only useful at that very last step. It's
not useful when you're trying to sign artifacts or stage them somewhere
and you don't care what the repo_name is, since they aren't going
through the final ship