(maint) Remove 'deb' from artifact path
This commit removes the 'deb' from the path of debian packages. We are
no longer staging debian packages under this directory. This is simply
an element I forgot to take out of the path. It wasn't causing any
errors because it was just getting subbed out at ship time. We shouldn't
be doing that either