(maint) Hard code path substitutions
Currently, we're setting msi_path, swix_path, and dmg_path to
'/opt/downloads/windows', '/opt/downloads/eos', and '/opt/downloads/mac'
respectively. We really shouldn't be handling these paths in that way.
These should be set to '/opt/downloads', and then the automation should
know to look in the respective directories. This matches with how all
the other platforms are handled. The problem is that we can't just
change these in the code bases. We need to be able to change them
gradually, and this is the best way to ensure backward compatibility. I
am so very sorry for writing this.