(maint) Update which latest dmg/msi symlinks we create
This commit changes which latest package symlinks we create. We
shouldn't have to maintain any of the old links (i.e., those for
puppet-3.x), as we are no longer shipping those packages. However, we do
still need to maintain the PC1 links, and the new puppet5 links. We can
do this in a slightly smarter way though.