(maint) Clean up linking command
Prior to this commit, the latest link command would attempt to create a
link to a target, even if that target didn't exist. This ended with
links to links, which was no fun. This commit cleans up the code so that
it simply skips the link if there is no target. We probably want to fail
if that is the case, but I don't think these links are be-all, end-all
  in terms of the infrastructure we have available.