(RE-8980) Drop sources into the correct directory
This commit updates the packaging repo to handle source packages
correctly. Previously, sources would just be dropped into the first
supported architecture for whatever platform-version we're dealing with.
This will let us send them to the proper directories when we ship.

This also adds a new method to allow us to get the package format
without needing to worry about architecture.