(RE-9133) Require a version be passed to release a package
This commit allows packages to be released by passing in a version to
the update_release_versions task instead of having packaging get this
package from the repo it is in. This needs to be done because tagging is
inconsistent across the various projects we have and our automation
should be able to pass versions down through jenkins jobs.