(maint) Do not report unversioned msis
Prior to this commit, we had been reporting unversioned msis in the
platform_data hash. This was fine for testing, as the unversioned msi is
simply a link to the versioned one. But this is bad when we go to ship
the packages. If we are fetching packages with FOSS_ONLY=true, it will
query the yaml data looking for foss only platform targets and fetch the
corresponding artifact. This is bad, since it means that for windows, we
were fetching the unversioned artifact for the final ship. This
changes ensures we don't set the unversioned msi when we are initially
writing out the platform_data hash.