(maint) Fix assumptions and typos
This commit fixes a few typos

It also fixes a bad assumption I had in the code. Previously, I was
assuming the path to the yaml file would be retrieved successfully based
on the existing path components. However, if we're trying to retrieve a
package, the yaml file isn't going to be under the same directories as
that package. We need a better way to figure out where to grab things.
The easiest way is to simply allow those values to be overwritten.