(RE-9525) Simultaneously ship nightlies to weth and old nightlies server
This commit adds an `uber_ship_lite` task that just retrieves builds
(ultimately from Artifactory) and ships them to the staging server. This will be
used to ship nightlies and will eventually include the tasks to update the yum
and apt repos and sync to s3 (once the script is updated for nightlies).