(maint) Only require artifactory gem if we are using artifactory
Previously, the fact that we had been requiring the artifactory gem at
the top of this file meant that calling any rake tasks would fail if the
artifactory gem was not installed. This commit moves the requires inside
the init for artifactory, so that we only require that library when we
are actually using it.