(maint) Make FOSS_ONLY work for nightlies
This commit allows the retrieve task to grab from the 'repos' directory, in
addition to the 'artifacts' directory, on builds.d.pl.net. Previously, this
restriction was in place because the <ref>.yaml file is necessary for FOSS_ONLY
retrieval, but the yaml file also exists in the 'repos' directory, so this
should work the same way.