(maint) Make artifacts_path optional for Pkg::Rpm::Repo.repo_creation_command
For the nightly repo ship and the internal s3 ship we aren't operating
on all of the packages staged on the distribution host, rather we're
operating on the packages staged locally. If the artifact_path isn't
specified, we'll run `createrepo` on every directory inside of the
`repo_directory` that contains rpms.