(RE-9889) Nightly repo ship only generates repo configs, ships no artifacts
This commit adds an optional parameter `legacy` to Pkg::Paths.repo_path.
Previously, the value of Pkg::Paths.repo_name, which switches based on whether
the project is a final version or not, was used to determine the repo_path. We
also recently changed the directory structure for rpms. This new param allows
the nightly ship task to find builds using the (final) repo name, rather than the
nonfinal repo name, when shipping to ravi and looks for repos using the old
directory structure when determinig what to ship. Ideally, this should be
reverted once the old nightlies process is obsoleted.