(RE-10247) Clojure project signing bundles don't have Gemfiles
This commit adds a check for Gemfile to the rake tasks that are executed
remotely. Previously, these commands would attempt a `bundle install` and then
fail if no Gemfile was found. Since not all projects are ready to use packaging
as a gem, we need to support the old method of using packaging (bootstrapping).