(maint) Add separate nightly repo update tasks
This commit adds rake tasks for updating nightly repos. Previously, we would
check whether the version of whatever we were shipping was final or not (based
on git describe) and switch on that, i.e. use nightly repo name, path, command,
etc. if nonfinal. This had a level of magic that easily allowed for accidental
repo updates in the wrong place.