(RE-10284) Separate final and nightly ship tasks
This commit adds separate rake tasks for shipping nightly packages. Previously,
we had single rake tasks for each package type that would switch between
shipping to the final repos and the nightly repos, based on the version of the
packages. This allows us to be intentional about what we're shipping where.