(maint) Update `sign_gem` to not rely on gem name and version
This commit changes the gem signing task to look for and sign all gems, not
just those that contain the gem name and version. This will allow for things
like the cfacter gem, which will get shipped to puppet-agent directories on
builds, to get shipped in the same process as shipping puppet-agent.