(maint) Remove `remote_target` param from `retrieve_all`
This commit removes the `remote_target` param from the `retrieve_all` function.
The `remote_target` would just get appended to the `build_url` and `rsync_path`
within the function, so now we just do the appending in the `retrieve` rake task
when setting the `build_url` and `build_path` there.