(maint) Add trailing slashes back into retrieve functions
This commit adds back the trailing '/' to `build_url` and `rsync_path`, used in
the `retrieve` functions. These use `wget` with the `--no-parent` flag: "for
HTTP (and HTTPS), the trailing slash is very important to ‘--no-parent’. HTTP
has no concept of a “directory”—Wget relies on you to indicate what’s a
directory and what isn’t. In ‘http://foo/bar/’, Wget will consider ‘bar’ to be a
directory, while in ‘http://foo/bar’ (no trailing slash), ‘bar’ will be
considered a filename (so ‘--no-parent’ would be meaningless, as its parent is
‘/’)."
(https://www.gnu.org/software/wget/manual/html_node/Directory_002dBased-Limits.html)
Adding a trailing slash to the rsync source path allows us to copy only the
directory's contents, rather than the whole directory itself.
(http://qdosmsq.dunbar-it.co.uk/blog/2013/02/rsync-to-slash-or-not-to-slash/)