(RE-10703) Add `PROJECT` environment variable
This commit adds a `PROJECT` environment variable that overrides/sets
`Pkg::Config.project` and adds explicit failures when shipping or retrieving
if the project is unset. Since we now have some repos that build multiple
different packages, we need to allow for different project names so that these
packages end up in separate directories on builds.