(maint) Bump rvm ruby version used in remote bundle install
This commit changes the ruby used when bundle installing on remote hosts from
2.3.1 to 2.4.1. Due to a rubygems bug
(https://github.com/rubygems/rubygems/issues/1448), bundler was attempting to
install the packaging gem in the wrong path. I was only able to reproduce this
issue using ruby 2.3.1, so I am bumping to the next later version that is
installed on the jenkins workers.