(PA-2016) Have package:gem noop if project doesn't build gem
This makes more sense. It is also important for PA-2016 where we want
to simplify our component artifact + shipping job to not rely on
jenkins:uber_build, b/c jenkins:uber_build dynamically generates several
jobs to do the work that makes it difficult to diagnose failures.

Our simplified job will just run all of the possible packaging tasks in
sequence. If the specific packaging task does not apply to the given
component, the task will noop and pass-thru. We ship all of the
generated artifacts at the end.