(maint) Bump minimum ruby version
This commit bumps the minimum required ruby version to 2.0.0. As we move toward
using the releng_metadata gem with packaging, this will be necessary since one
of releng_metadata's transitive dependencies requires ruby >= 2.0. 1.9.3 is also
just really old version and we already removed ruby 1.9.3 travis test matrix, so
this finishes the job.