(maint) Loosen rake dependency
Looking at the blame from when the original pinning was done, I could not find
any info for *why* we pinned to 12.2.1. I assume it was because rake 2.3.x
drops support for ruby 1.9.x. Now that we are also dropping support for 1.9.x,
we should be able to safely use later versions of rake.