(RE-10924) Refactor `has_sig?` method
This commit refactors the `has_sig?` method to use more ruby functionality
instead of shelling out to do all the work. This allows us to test this method,
which this commit also includes. Finally, `has_sig?` now checks for the short
gpg key in the `rpm -Kv` output, rather than the full key.