(maint) Don't use `capture3` in `has_sig?`
This commit replaces the use of `capture3` with `%x`. When checking signatures,
the `rpm -Kv` command will exit 1 if the signing key hasn't been loaded as a
trusted key, even if the package is signed. Since all we care about is whether
the package is signed and since `capture3` bails out on non-0 exits, we have to
use something else.