(maint) Export PACKAGING_LOCATION before doing remote bundle installs (take two)
Previously, even when PACKAGING_LOCATION is set locally, remote tasks (like
signing) were still pulling in the default version of packaging. Now, we pass
along this environment variable as part of our remote commands. Note that this
will not work if PACKAGING_LOCATION is set to a local directory. Instead, use
a git branch.