(maint) rsync back to root_dir after signing
This commit updates the `sign_all` task to rsync the signed packages back to
the `root_dir` that the packages were originally in, before signing.
Previously, we rsync'd back to `DEFAULT_DIRECTORY`, which may or may not be the
same as `root_dir`. This was an issue when signing vanagon projects: builds are
staged in a directory called 'output' and then shipped (internally) from there;
the `DEFAULT_DIRECTORY` is 'pkg', so when we attemped to sign between building
and shipping (internally), signed packages were getting rsync'd to 'pkg' but
the unsigned packages in 'output' were getting shipped.