(maint) Add quotes around PACKAGING_LOCATION value when exporting
This commit adds single quotes around the value of ENV['PACKAGING_LOCATION']
when it is exported as part of the remote bundle install process. Previously,
if the value had spaces in it (e.g. git://github.com/mwaggett/packaging.git
--branch=foo), everything after the space would get left out of the export and
thing would start failing, since the remainder is not a valid command.