(RE-11480) Add `ship_nightly_gem` task
This commit adds a rake task for shipping nightly gems. This task will ship the
gems to Artifactory and to the nightly repos on weth. This commit also adds a
`nonfinal_gem_path` param, similar to existing `nonfinal_<file type>_path`
params, to distinguish between the final `gem_path` - this param will need to
be added to build_defaults and/or build-data for any projects that wish to ship
nightly gems.