(maint) Use `internal_gem_host` param in `ship_to_internal_mirror` method
This commit replaces the hard-coded artifactory url in the
`ship_to_internal_mirror` method with the existing `internal_gem_host` param.
This commit also edits the value of this param to not have a trailing slash
because artifactory urls are weird and `gem push`ing fails if the artifactory
url has a trailing slash. This param doesn't appear to actually be in use
anywhere in this repo, so this change shouldn't affect anything.