(RE-11600) Skip signing msis if they're already signed
This commit adds a conditional to the msi signing method to skip signing if the
msi is already signed. As we work to move signing up in our build process, we
want to ensure that artifacts are signed only once so that all copies of the
artifact have the same checksum.