(RE-11600) Skip signing p5ps if they're already signed
This commit adds a block to the ips signing method to check if the p5p file is
signed already before proceeding with repo creation and signing. As we work to
move signing up in our build process, we want to ensure that artifacts are
signed only once so that all copies of the artifact have the same checksum.