(PE-16465) Update logic to create the "all the agents" tarball
When the surrounding logic changed to use the yaml config for
generating the list of platforms (it used to be hardcoded in the job),
the existing code that generates the tarball containing all of the
agents broke. This moves the bulk of the logic out of the rake task
and into the library, and updates to work with the newer layout.