(maint) Add method for determining release package link paths
This commit adds a `release_package_link_path` method that takes a platform tag
and constructs the the target path for the top-level release package symlink
associated with that platform, taking into account whether this is a final or
nonfinal (i.e. nightly) release package.
Currently, updating these symlinks is a manual process. This is the first step
in an effort to automate these updates.