(maint) Sign deb changes files individually
This commit updates the `sign_deb_changes` task to loop through the .changes
files and sign them individually instead of passing a glob to the sign method.
Previously, this task was probably never getting called, since our shipping
workflow doesn't retrieve .changes files. Now that we have moved signing to
build time, we encountered an issue where the glob was not being interpretted /
expanded properly. This aims to fix that issue.