(RE-11291) Improve rsync error message
This commit updates the error message generated when attempting to rsync while
both origin_host and target_host are nil. This most-often occurs when the TEAM
environment variable is unset, since params from build_data#release are not
fetched in this case. The new message provides more clarity into the nature of
the error and how to resolve it.