(RE-11584) Replace `remote_bootstrap` method with `remote_unpack_git_bundle`
This commit replaces the `remote_bootstrap` method with a new
`remote_unpack_git_bundle` method that does everything `remote_bootstrap`
did except actually bootstrapping or bundle installing. The
`remote_bundle_install_command` is now called as part of the same command that
calls the remote rake tasks. Previously, even though bundle installing appeared
to be successful when called in a separate command, the rake command would fail
because it could not find the packaging gem. By calling the `bundle install` in
the same command as the rake call, we ensure that we're in the same bundle
environment and all gems are accessible.