(RE-11741) Narrow search for latest package symlink creation
This commit updates the command used to determine the latest package for a
given project in a given repository. Previously, we just grepped for the
project name of whatever we're currently shipping. This resulted in
'puppet-latest' symlinks getting created in the windows/ and mac/ directories,
whenever we ship puppet, even though we don't ship puppet msis or dmg, since
puppet-agent and puppet-bolt both match 'puppet'.
Now we grep -P for '<project_name>-\d', which assumes package names of the form
'<project_name>-<version>', to prevent this erroneous matching.