(RE-11984) Use quotes when grepping a regex
This commit escapes the '\d' and adds quotes around the <package_name>-<version>
regex we use for finding the latest version of a project and symlinking to it.
Previously, without the escape or quotes, we ended up searching for the literal
string '-d' after the package name (e.g. 'puppet-bolt-d'), which resulted in no
new -latest symlinks when we shipped a new version.