(RE-11887) Check if gem has already been shipped before shipping
This commit adds a `shipped_to_rubygems?` method that checks whether a given
gem version exists already on rubygems.org. This method is then used in the
`ship_to_rubygems` method to skip the ship attempt if the gem has already been
shipped. Previously, `ship_to_rubygems` would fail if we tried to ship an
already-shipped gem. We don't want this method to fail because as we automate
the process of shipping puppet-agent and its components, we would like to be
able to just look at puppet-agent's component configs for the appropriate
component version without having to know if that version is new or if it's the
same version that was in the previous release. Now we can attempt to ship an
already-shipped gem without things blowing up.