(maint) Rescue JSON parse error if gem isn't found
This commit updates the `shipped_to_rubygems?` method to rescue errors and
return false if the given gem isn't found anywhere on rubygems yet. We assume
that this would happen if we're shipping a new gem for the first time, so we
don't want to fail. Instead, by returning false, we allow the attempt to ship
to proceed.