(maint) Fixes for `retrieve_yaml_data`
The `retrieve_yaml_data` method was operating under the assumption that
each yaml file name would be unique. While that's the case for SHAs,
for tagged versions this won't work. The method now requires the package
name and git ref to be passed.

This also removes support for retrieving packages without passing a
package name from the `retrieve_package` method. That functionality
wasn't working, isn't in use anywhere, and doesn't make sense if we're
requiring package name and ref to locate the YAML.