(CPR-677) Create 'latest' symlinks for nightly msis and dmgs
This commit updates the calls to the `remote_create_latest_symlink` method to
take into account whether we are shipping nightlies or not. Previously, the
path we tried to link from for nightlies included a final repo name (e.g.
puppet6 instead of puppet6-nightly), so those paths weren't found in the
nightly repositories and links were not created.