(CPR-698) Update `stage_apt_archives` method
This commit updates the `stage_apt_archives` method to move freight packages
to a staging directory where repos will be created and move apt packages to
a temporary directory where they will ultimately be deleted. Previously, when
we thought we wern't going to create archive repos, this method moved the apt
packages to a staging directory for syncing and moved the freight packages to
a temporary directory for deletion.