(maint) Make the Pkg::Deb::Repo #sign_repos spec allow for the 'keychain' command.
In an attempt into understanding the SIGTERM issues we've been
occasionally spotting in packaging, I noticed that rspec didn't run on
my machine. Turns out I had 'keychain' installed and the spec tests
didn't allow for that. rspec was failing on my box.

Update the spec test to allow for this possibility.
Also sneak in a minor bit of code refactoring in Pkg::Deb::Repo.sign_repos