(maint) Fail shipping gems to downloads if gem_path is unset
This commit adds a check for the `gem_path` parameter when attempting to ship
gems to downloads.puppet.com. Without this, either the job will fail anyway
later with an unset param or it will ship to an undesired path (I didn't
check this behavior, but either way seems Bad). This allows us to be explicit
about where we ship gems.