(RE-12793) Allow rake versions >=13.0.0 to be used at runtime
Rake 13.0.0 was released on September 28, 2019. Because of rake required
as a runtime requirement, using '~> 12.3' as the Rake dependency was
causing Bundler to return back to a much earlier version of packaging.

The removes that restriction and allows packaging to run with a
more relaxed Rake version requirement.

This could bite us in the future. Ideally, we should not be using Rake
at runtime.