(RE-12788) Set `deb.component` property when copying PE packages
This commit sets the `deb.component` property on copied debian packages when
populating a new PE repo. We set this to the PE version, which allows the
versioned subdirectories in Artifactory to be treated as separate repository
components. Without this, packages won't show up in the `dists` directory and
repo metadata won't pick them up, meaning we can't download the packages via
package manager.