(RE-12869) Refactor `ship_gem_to_downloads` task
This commit reorganizes the `ship_gem_to_downloads` task to be clearer about
the cases where we do and don't want to actually attemp shipping to downloads.
Previously, the `exit` that was called to prevent the upload in the case of
unset params was preventing downstream tasks from getting run. For example,
because Bolt ships gems to rubygems, but not downloads, they have `build_gem`
set to true but no `gem_host` or `gem_path`. When shipping, none of the
`uber_ship_lite` tasks after `ship_gem_to_downloads` were getting run.