(maint) Avoid race conditions pushing gem to rubygems
In order to avoid re-pushing to rubygems if we have already pushed a version we check the `versions` api. There appears to be a race condition where that API has out-dated infor by the time we go to push. This commit catches the "gem already pushed" error, logs it and will *not* throw it as the intent of the action is already satisfied.