Export GEM_SOURCE to remote machine
Previously, when executing a command on a remote machine, packaging
exported some environmental variables such as PACKAGING_LOCATION and
VANAGON_LOCATION which allowed commands run on a remote machine to
maintain the same environment.

However, GEM_SOURCE was not among those exported environmental
variables. If a project was using a gem not found on Rubygems--in my
case, pl_fustigit hosted on Artifactory--that gem could be installed
locally by using the source defined in GEM_SOURCE, but not remotely.

This commit checks whether the GEM_SOURCE environmental variable is
defined and, if it is and isn't empty, exports it to a remote machine
with other environmental variables.