(PE-33254) Move solaris signing to GPG
Solaris .p5p/.ips signing no longer works because Digicert no longer
supports disk-bound root CAs and one is obliged to use their
utilities to retrieve the Root CA for signing.

New-style Retrieved root CAs are time-bound and can't readily be
re-used. Thus they need to be re-fetched every time a signing is needed.

Digicert does not provide a Solaris version of their CA-fetching tool.

A limpy way past this is to provide a GPG-based .asc file which
lives next to the .p5p which users can use for verification provided
that they've chosen to trust our signing key.

We use this method for a few other packages (not the least of which
are PE tarballs)